+++ terminal/backends/__init__.py
@@ -1,5 +1,12 @@
+from .pipe_backend import PipeProcessSession
 from .pty_backend import PtyProcessSession, UnsupportedPtySession
 from .tmux_backend import TmuxSession
 from .winpty_backend import WinPtySession
 
-__all__ = ["PtyProcessSession", "TmuxSession", "UnsupportedPtySession", "WinPtySession"]
+__all__ = [
+    "PipeProcessSession",
+    "PtyProcessSession",
+    "TmuxSession",
+    "UnsupportedPtySession",
+    "WinPtySession",
+]